MainScreen.kt and its tests have been implemented. Fixes: type of timestamp in StockUi.kt changed to long echoJob isActive check issue fixed at EchoStockRepositoryImpl.kt StockIndicator logic implemented to StockViewModel.kt (Will be moved to repository or usecase)

+++ app/src/main/java/com/ramazanm/rtpricetrackerdemo/data/model/StockUi.kt
@@ -3,7 +3,7 @@ package com.ramazanm.rtpricetrackerdemo.data.model
 data class StockUi(
     val name: String = "",
     val price: Double = 0.0,
-    val timestamp: Int = 0,
+    val timestamp: Long = 0,
     val indicator: StockIndicator = StockIndicator.NEUTRAL
 )
